+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.1-dev.3
+
+- **BREAKING** refactor: rename `positionalArgs` to `positionalArguments`
+- **BREAKING** refactor: rename `namedArgs` to `namedArguments`
+- fix: `thenAnswer` callback contains correct `Invocation`
+- docs: minor documentation updates
+
 # 0.0.1-dev.2
 
 - docs: minor updates to README